move various testrunners into their own file.
