+++ js/client/modules/@arangodb/testutils/test-utils.js
@@ -28,19 +28,9 @@
 /* Modules: */
 const _ = require('lodash');
 const fs = require('fs');
-const pu = require('@arangodb/testutils/process-utils');
-const ct = require('@arangodb/testutils/client-tools');
-const yaml = require('js-yaml');
 
-const toArgv = require('internal').toArgv;
-const time = require('internal').time;
-const sleep = require('internal').sleep;
-const download = require('internal').download;
 const pathForTesting = require('internal').pathForTesting;
 const platform = require('internal').platform;
-const SetGlobalExecutionDeadlineTo = require('internal').SetGlobalExecutionDeadlineTo;
-const userManager = require("@arangodb/users");
-const testRunnerBase = require('@arangodb/testutils/testrunner').testRunner;
 const setDidSplitBuckets = require('@arangodb/testutils/testrunner').setDidSplitBuckets;
 const isEnterprise = require("@arangodb/test-helper").isEnterprise;
 
@@ -324,277 +314,10 @@ function scanTestPaths (paths, options, fun) {
 }
 
 
-function getTestCode(file, options, instanceManager) {
-  let filter;
-  if (options.testCase) {
-    filter = JSON.stringify(options.testCase);
-  } else if (options.failed) {
-    let failed = options.failed[file] || options.failed;
-    filter = JSON.stringify(Object.keys(failed));
-  }
-
-  let runTest;
-  if (file.indexOf('-spec') === -1) {
-    filter = filter || '"undefined"';
-    runTest = 'const runTest = require("jsunity").runTest;\n';
-
-  } else {
-    filter = filter || '';
-    runTest = 'const runTest = require("@arangodb/mocha-runner");\n';
-  }
-  let ret = '';
-  if (instanceManager != null) {
-    ret = 'global.instanceManager = ' + JSON.stringify(instanceManager.getStructure()) + ';\n';
-  }
-  return ret + runTest + 'return runTest(' + JSON.stringify(file) + ', true, ' + filter + ');\n';
-}
-// //////////////////////////////////////////////////////////////////////////////
-// / @brief runs a remote unittest file using /_admin/execute
-// //////////////////////////////////////////////////////////////////////////////
-
-class runOnArangodRunner extends testRunnerBase{
-  constructor(options, testname, ...optionalArgs) {
-    super(options, testname, ...optionalArgs);
-    this.info = "onRemoteArangod";
-  }
-  runOneTest(file) {
-    try {
-      let testCode = getTestCode(file, this.options, this.instanceManager);
-      let httpOptions = _.clone(this.instanceManager.httpAuthOptions);
-      httpOptions.method = 'POST';
-
-      httpOptions.timeout = this.options.oneTestTimeout;
-      if (this.options.isSan) {
-        httpOptions.timeout *= 2;
-      }
-      if (this.options.valgrind) {
-        httpOptions.timeout *= 2;
-      }
-
-      httpOptions.returnBodyOnError = true;
-      const reply = download(this.instanceManager.url + '/_admin/execute?returnAsJSON=true',
-                             testCode,
-                             httpOptions);
-      if (!reply.error && reply.code === 200) {
-        return JSON.parse(reply.body);
-      } else {
-        if ((reply.code === 500) &&
-            reply.hasOwnProperty('message') &&
-            (
-              (reply.message.search('Request timeout reached') >= 0 ) ||
-                (reply.message.search('timeout during read') >= 0 ) ||
-                (reply.message.search('Connection closed by remote') >= 0 )
-            )) {
-          print(RED + Date() + " request timeout reached (" + reply.message +
-                "), aborting test execution" + RESET);
-          return {
-            status: false,
-            message: reply.message,
-            forceTerminate: true
-          };
-        } else if (reply.hasOwnProperty('body')) {
-          return {
-            status: false,
-            message: reply.body
-          };
-        } else {
-          return {
-            status: false,
-            message: yaml.safeDump(reply)
-          };
-        }
-      }
-    } catch (ex) {
-      return {
-        status: false,
-        message: ex.message || String(ex),
-        stack: ex.stack
-      };
-    }
-  }
-}
-
-function readTestResult(path, rc, testCase) {
-  const jsonFN = fs.join(path, 'testresult.json');
-  let buf;
-  try {
-    buf = fs.read(jsonFN);
-    fs.remove(jsonFN);
-  } catch (x) {
-    let msg = 'readTestResult: failed to read ' + jsonFN + " - " + x;
-    print(RED + msg + RESET);
-    rc.message += " - " + msg;
-    rc.status = false;
-    return rc;
-  }
-
-  let result;
-  try {
-    result = JSON.parse(buf);
-  } catch (x) {
-    let msg = 'readTestResult: failed to parse ' + jsonFN + "'" + buf + "' - " + x;
-    print(RED + msg + RESET);
-    rc.message += " - " + msg;
-    rc.status = false;
-    return rc;
-  }
-
-  if (Array.isArray(result)) {
-    if (result.length === 0) {
-      // spec-files - don't have parseable results.
-      rc.failed = rc.status ? 0 : 1;
-      return rc;
-    } else if ((result.length >= 1) &&
-               (typeof result[0] === 'object') &&
-               result[0].hasOwnProperty('status')) {
-      return result[0];
-    } else {
-      rc.failed = rc.status ? 0 : 1;
-      rc.message = "don't know howto handle '" + buf + "'";
-      return rc;
-    }
-  } else if (_.isObject(result)) {
-    if ((testCase !== undefined) && result.hasOwnProperty(testCase)) {
-      return result[testCase];
-    } else {
-      if (rc.hasOwnProperty('exitCode') && rc.exitCode !== 0) {
-        result.failed += 1;
-        result.status = false;
-        result.message = rc.message;
-      }
-      return result;
-    }
-  } else {
-    rc.failed = rc.status ? 0 : 1;
-    rc.message = "readTestResult: don't know howto handle '" + buf + "'";
-    return rc;
-  }
-}
-
-function writeTestResult(path, data) {
-  const jsonFN = fs.join(path, 'testresult.json');
-  fs.write(jsonFN, JSON.stringify(data));
-}
-
-
-// //////////////////////////////////////////////////////////////////////////////
-// / @brief runs a local unittest file using arangosh
-// //////////////////////////////////////////////////////////////////////////////
-
-class runInArangoshRunner extends testRunnerBase {
-  constructor(options, testname, ...optionalArgs) {
-    super(options, testname, ...optionalArgs);
-    this.info = "forkedArangosh";
-  }
-  getEndpoint() {
-    return this.instanceManager.findEndpoint();
-  }
-  runOneTest(file) {
-    require('internal').env.INSTANCEINFO = JSON.stringify(this.instanceManager.getStructure());
-    let args = ct.makeArgs.arangosh(this.options);
-    args['server.endpoint'] = this.getEndpoint();
-
-    args['javascript.unit-tests'] = fs.join(pu.TOP_DIR, file);
-
-    args['javascript.unit-test-filter'] = this.options.testCase;
-
-    if (this.options.forceJson) {
-      args['server.force-json'] = true;
-    }
-
-    if (!this.options.verbose) {
-      args['log.level'] = 'warning';
-    }
-    if (this.options.extremeVerbosity === true) {
-      args['log.level'] = 'v8=debug';
-    }
-    if (this.addArgs !== undefined) {
-      args = Object.assign(args, this.addArgs);
-    }
-    // TODO require('internal').env.INSTANCEINFO = JSON.stringify(this.instanceManager);
-    let rc = pu.executeAndWait(pu.ARANGOSH_BIN, toArgv(args), this.options, 'arangosh', this.instanceManager.rootDir, this.options.coreCheck);
-    return readTestResult(this.instanceManager.rootDir, rc, args['javascript.unit-tests']);
-  }
-}
-
-
-// //////////////////////////////////////////////////////////////////////////////
-// / @brief runs a local unittest file in the current arangosh
-// //////////////////////////////////////////////////////////////////////////////
-
-class runLocalInArangoshRunner extends testRunnerBase {
-  constructor(options, testname, ...optionalArgs) {
-    super(options, testname, ...optionalArgs);
-    this.info = "localArangosh";
-  }
-  runOneTest(file) {
-    let endpoint = arango.getEndpoint();
-    if (this.options.vst || this.options.http2) {
-      let newEndpoint = this.instanceManager.findEndpoint();
-      if (endpoint !== newEndpoint) {
-        print(`runInLocalArangosh: Reconnecting to ${newEndpoint} from ${endpoint}`);
-        arango.reconnect(newEndpoint, '_system', 'root', '');
-      }
-    }
-
-    let testCode = getTestCode(file, this.options, null);
-    global.instanceManager = this.instanceManager;
-    let testFunc;
-    try {
-      eval('testFunc = function () {\n' + testCode + "}");
-    } catch (ex) {
-      print(RED + 'test failed to parse:');
-      print(ex);
-      print(RESET);
-      return {
-        status: false,
-        message: "test doesn't parse! '" + file + "' - " + ex.message || String(ex),
-        stack: ex.stack
-      };
-    }
-
-    try {
-      SetGlobalExecutionDeadlineTo(this.options.oneTestTimeout * 1000);
-      let result = testFunc();
-      let timeout = SetGlobalExecutionDeadlineTo(0.0);
-      if (timeout) {
-        return {
-          timeout: true,
-          forceTerminate: true,
-          status: false,
-          message: `test aborted due to ${require('internal').getDeadlineReasonString()}. Original test status: ${JSON.stringify(result)}`,
-        };
-      }
-      if (result === undefined) {
-        return {
-          timeout: true,
-          status: false,
-          message: "test didn't return any result at all!"
-        };
-      }
-      return result;
-    } catch (ex) {
-      let timeout = SetGlobalExecutionDeadlineTo(0.0);
-      print(RED + 'test has thrown: ' + (timeout? "because of timeout in execution":""));
-      print(ex, ex.stack);
-      print(RESET);
-      return {
-        timeout: timeout,
-        forceTerminate: true,
-        status: false,
-        message: "test has thrown! '" + file + "' - " + ex.message || String(ex),
-        stack: ex.stack
-      };
-    }
-  }
-}
 
 exports.testServerAuthInfo = testServerAuthInfo;
 exports.testClientJwtAuthInfo = testClientJwtAuthInfo;
 
-exports.runOnArangodRunner = runOnArangodRunner;
-exports.runInArangoshRunner = runInArangoshRunner;
-exports.runLocalInArangoshRunner = runLocalInArangoshRunner;
 
 exports.makePathUnix = makePathUnix;
 exports.makePathGeneric = makePathGeneric;
